#7: Account for click bonus for quantities of products needed at the top (#41)
When choosing this option, it is assumed that the top product and its up
to two inputs are being clicked every 150ms or so, making 3.5x as many
cycles as would be typical, so it adjusts the required inputs by that
quantity. It is also assumed that the click boost is active, so adjusts
the outputs by that quantity. Autoclickers FTW.